# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IAEDU API",
        "description": "API para el sistema educativo IAEDU",
        "contact": {
            "email": "admin@iaedu.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Obtener lista de calificaciones",
                "operationId": "014720d431d750512c5cb811b567e605",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Filtrar por estudiante",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "description": "Filtrar por materia",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Filtrar por grupo",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Incluir relaciones (student,subject)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de calificaciones obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Grades retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Grade"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Crear nueva calificación",
                "operationId": "3cc5e1c1b6111ce57dd28b630adb2812",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "subject_id",
                                    "promedio_final"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subject_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "promedio_final": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.5
                                    },
                                    "evaluations": {
                                        "type": "object"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "Aprobado"
                                    },
                                    "faltantes": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "puntos_faltantes": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Calificación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Grade created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Grade"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/grades/{id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Obtener calificación específica",
                "operationId": "2a9dbd1af4ae65916b81af86e65445af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la calificación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "Incluir relaciones (student,subject)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calificación obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Grade retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Grade"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Calificación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Grades"
                ],
                "summary": "Actualizar calificación",
                "operationId": "2c7326238b573c059e38817a80acef69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la calificación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subject_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "promedio_final": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.5
                                    },
                                    "evaluations": {
                                        "type": "object"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "Aprobado"
                                    },
                                    "faltantes": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "puntos_faltantes": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calificación actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Grade updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Grade"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Calificación no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Eliminar calificación",
                "operationId": "28a518d3c74a7de53f0c33d0331b5021",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la calificación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calificación eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Grade deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Calificación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Grade": {
                "title": "Grade",
                "description": "Modelo de calificación del sistema",
                "properties": {
                    "id": {
                        "description": "ID único de la calificación",
                        "type": "integer",
                        "example": 1
                    },
                    "student_id": {
                        "description": "ID del estudiante",
                        "type": "integer",
                        "example": 1
                    },
                    "subject_id": {
                        "description": "ID de la materia",
                        "type": "integer",
                        "example": 1
                    },
                    "promedio_final": {
                        "description": "Promedio final de la calificación",
                        "type": "number",
                        "format": "float",
                        "example": 8.5
                    },
                    "evaluations": {
                        "description": "Evaluaciones individuales",
                        "type": "object"
                    },
                    "estado": {
                        "description": "Estado de la calificación",
                        "type": "string",
                        "example": "Aprobado"
                    },
                    "faltantes": {
                        "description": "Número de evaluaciones faltantes",
                        "type": "integer",
                        "example": 0
                    },
                    "puntos_faltantes": {
                        "description": "Puntos faltantes",
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "date": {
                        "description": "Fecha de la calificación",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}